resolve cyclic imports by pulling out the abstract BLL
